Fixing not removing checkpoints from the queue of pending checkpoints.
Once we were downloading a checkpoint, it was never removed from the list of downloading checkpoints.
This would cause a conflict if we are attempting to upload a checkpoint that is actually
a rollback image, rather than a standard image. This would cause the rollback checkpoint to fail to be
uploaded.

Its not clear if this is a problem with the new implemenetation, or something that can always
exist in the previous implementation... say there are multiple standby checkpointers running,